Make logo link to overview
