+++ packages/experience-docs/source/components/App/index.js
@@ -1,24 +1,40 @@
 import React from 'react'
+import find from 'lodash/find'
+import trimEnd from 'lodash/trimEnd'
 
 import PageWrapper from '../PageWrapper'
 import Navigation from '../Navigation'
 import Content from '../Content'
+import Header from '../Header'
 import ContentWrapper from '../ContentWrapper'
 import FeedbackLink from '../FeedbackLink'
 
-import content from '../../content'
+import content, { flatContent } from '../../content'
 
 const App = ({
   children,
   location
-}) => (
-  <PageWrapper>
-    <ContentWrapper>
-      <Content activeRoute={location.pathname}>{children}</Content>
-      <FeedbackLink />
-    </ContentWrapper>
-    <Navigation routes={content} activeRoute={location.pathname} />
-  </PageWrapper>
-)
+}) => {
+  const currentPage = find(flatContent, (route) =>
+    route.attributes &&
+    trimEnd(route.attributes.path, '/') === trimEnd(location.pathname, '/'))
+  const status = currentPage && currentPage.attributes
+    ? currentPage.attributes.status
+    : 'red'
+  const title = currentPage && currentPage.attributes
+    ? currentPage.attributes.title : 'Missing Title'
+  return (
+    <PageWrapper>
+      <ContentWrapper>
+        <Header status={status} />
+        <Content title={title}>
+          {children}
+        </Content>
+        <FeedbackLink />
+      </ContentWrapper>
+      <Navigation routes={content} activeRoute={location.pathname} />
+    </PageWrapper>
+  )
+}
 
 export default App
